refactor: extract common viem provider/signer logic into shared base classes
